Fixed bugs and added interpreter features
cond had bugs, along with other bugs elsewhere that were fixed
Interpreter can now take in a file as input and run it.
More primitives added
Also added better error returns, as objects with a string and a type of error.
Overall, interpreter is more robust to errors, leaving error handling to the interpreter program itself(throwing the error upwards)